requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
arch==7.0.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
catboost==1.2.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cvxpy==1.5.2
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==5.0.1
djangorestframework==3.15.1
dnspython==2.6.1
docopt==0.6.2
ecos==2.0.14
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.1.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.1
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.1
keras==3.3.3
kiwisolver==1.4.5
langchain==0.2.11
langchain-core==0.2.23
langchain-text-splitters==0.2.2
langsmith==0.1.93
libclang==18.1.1
llvmlite==0.42.0
loguru==0.7.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mlpack==4.3.0.post1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
namex==0.0.8
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.2.4
numba==0.59.1
numpy==1.26.3
ollama==0.2.1
openai==1.37.0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.6
osqp==0.6.7.post1
packaging==23.2
pandas==2.2.0
pandas-datareader==0.10.0
parso==0.8.4
patsy==0.5.6
peewee==3.17.1
pexpect==4.9.0
pi==0.1.2
pillow==10.2.0
pipwin==0.5.2
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.5.0
preprocessing==0.1.13
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycares==4.4.0
pycodestyle==2.12.0
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.17.2
pyjsparser==2.7.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyportfolioopt==1.5.5
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==8.3.1
pythclient==0.1.24
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.0.3
qdldl==0.1.7.post4
radian==0.6.10
rchitect==0.4.6
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rpy2==3.5.16
scikit-learn==1.4.2
scipy==1.12.0
scs==3.2.6
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sphinx-rtd-theme==0.2.4
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.34.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.12.3
typing_extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.30.3
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.37
